+++ cbc/notes_test.go
@@ -23,3 +23,26 @@ func TestNotesValidation(t *testing.T) {
 	assert.Error(t, err)
 	assert.Contains(t, err.Error(), "key: must be a valid value")
 }
+
+func TestNoteSameAs(t *testing.T) {
+	n := &cbc.Note{
+		Key:  cbc.NoteKeyGeneral,
+		Code: "ABC",
+		Text: "This is a test",
+	}
+	assert.True(t, n.SameAs(&cbc.Note{
+		Key:  cbc.NoteKeyGeneral,
+		Code: "ABC",
+		Text: "This is a test ABC",
+	}))
+	assert.False(t, n.SameAs(&cbc.Note{
+		Key:  cbc.NoteKeyGeneral,
+		Code: "123",
+		Text: "This is a test 123",
+	}))
+	assert.False(t, n.SameAs(&cbc.Note{
+		Key:  cbc.NoteKeyLegal,
+		Code: "ABC",
+		Text: "This is a test ABC",
+	}))
+}