+++ regimes/it/categories.go
@@ -5,6 +5,7 @@ import (
 	"github.com/invopop/gobl/i18n"
 	"github.com/invopop/gobl/num"
 	"github.com/invopop/gobl/tax"
+	"github.com/invopop/validation"
 )
 
 // Local tax category definitions which are not considered standard.
@@ -30,6 +31,17 @@ var categories = []*tax.Category{
 		Title: i18n.String{
 			i18n.EN: "Value Added Tax",
 			i18n.IT: "Imposta sul Valore Aggiunto",
+		},
+		Validation: func(c *tax.Combo) error {
+			return validation.ValidateStruct(c,
+				validation.Field(&c.Ext,
+					validation.When(
+						c.Percent == nil,
+						tax.ExtensionsRequires(ExtKeySDIExempt),
+					),
+					validation.Skip,
+				),
+			)
 		},
 		Rates: []*tax.Rate{
 			{
@@ -100,7 +112,7 @@ var categories = []*tax.Category{
 					i18n.IT: "Esente",
 				},
 				Extensions: []cbc.Key{
-					ExtKeySDINature,
+					ExtKeySDIExempt,
 				},
 			},
 		},
@@ -118,10 +130,11 @@ var categories = []*tax.Category{
 			i18n.EN: "Personal Income Tax",
 			i18n.IT: "Imposta sul Reddito delle Persone Fisiche",
 		},
+		Validation: requireRetainedReason,
 		Map: cbc.CodeMap{
 			KeyFatturaPATipoRitenuta: "RT01",
 		},
-		Extensions: []cbc.Key{ExtKeySDIRetainedTax},
+		Extensions: []cbc.Key{ExtKeySDIRetained},
 	},
 	{
 		Code:     TaxCategoryIRES,
@@ -134,10 +147,11 @@ var categories = []*tax.Category{
 			i18n.EN: "Corporate Income Tax",
 			i18n.IT: "Imposta sul Reddito delle Società",
 		},
+		Validation: requireRetainedReason,
 		Map: cbc.CodeMap{
 			KeyFatturaPATipoRitenuta: "RT02",
 		},
-		Extensions: []cbc.Key{ExtKeySDIRetainedTax},
+		Extensions: []cbc.Key{ExtKeySDIRetained},
 	},
 	{
 		Code:     TaxCategoryINPS,
@@ -150,7 +164,8 @@ var categories = []*tax.Category{
 			i18n.EN: "Contribution to the National Social Security Institute",
 			i18n.IT: "Contributo Istituto Nazionale della Previdenza Sociale", // nolint:misspell
 		},
-		Extensions: []cbc.Key{ExtKeySDIRetainedTax},
+		Validation: requireRetainedReason,
+		Extensions: []cbc.Key{ExtKeySDIRetained},
 		Map: cbc.CodeMap{
 			KeyFatturaPATipoRitenuta: "RT03",
 		},
@@ -166,7 +181,8 @@ var categories = []*tax.Category{
 			i18n.EN: "Contribution to the National Welfare Board for Sales Agents and Representatives",
 			i18n.IT: "Contributo Ente Nazionale Assistenza Agenti e Rappresentanti di Commercio", // nolint:misspell
 		},
-		Extensions: []cbc.Key{ExtKeySDIRetainedTax},
+		Validation: requireRetainedReason,
+		Extensions: []cbc.Key{ExtKeySDIRetained},
 		Map: cbc.CodeMap{
 			KeyFatturaPATipoRitenuta: "RT04",
 		},
@@ -182,9 +198,19 @@ var categories = []*tax.Category{
 			i18n.EN: "Contribution to the National Pension and Welfare Board for Doctors",
 			i18n.IT: "Contributo - Ente Nazionale Previdenza e Assistenza Medici", // nolint:misspell
 		},
-		Extensions: []cbc.Key{ExtKeySDIRetainedTax},
+		Validation: requireRetainedReason,
+		Extensions: []cbc.Key{ExtKeySDIRetained},
 		Map: cbc.CodeMap{
 			KeyFatturaPATipoRitenuta: "RT05",
 		},
 	},
+}
+
+func requireRetainedReason(c *tax.Combo) error {
+	return validation.ValidateStruct(c,
+		validation.Field(&c.Ext,
+			tax.ExtensionsRequires(ExtKeySDIRetained),
+			validation.Skip,
+		),
+	)
 }