+++ tax/scenario.go
@@ -52,17 +52,18 @@ type Scenario struct {
 	// situations.
 	Codes cbc.CodeMap `json:"codes,omitempty" jsonschema:"title=Codes"`
 
-	// Any additional local meta data that may be useful in integrations.
-	Meta cbc.Meta `json:"meta,omitempty" jsonschema:"title=Meta"`
+	// Ext represents a set of tax extensions that should be applied to
+	// the document in the appropriate "tax" context.
+	Ext Extensions `json:"ext,omitempty" jsonschema:"title=Extensions"`
 }
 
 // ScenarioSummary is the result after running through a set of
-// scenarios and determining which combinations of Notes and Meta
-// are viable.
+// scenarios and determining which combinations of Notes, Codes, Meta,
+// and extensions are viable.
 type ScenarioSummary struct {
 	Notes []*cbc.Note
 	Codes cbc.CodeMap
-	Meta  cbc.Meta
+	Ext   Extensions
 }
 
 // ValidateWithContext checks the scenario set for errors.
@@ -80,7 +81,7 @@ func (ss *ScenarioSet) SummaryFor(docType cbc.Key, docTags []cbc.Key, docExt []E
 	summary := &ScenarioSummary{
 		Notes: make([]*cbc.Note, 0),
 		Codes: make(cbc.CodeMap),
-		Meta:  make(cbc.Meta),
+		Ext:   make(Extensions),
 	}
 	for _, s := range ss.List {
 		if s.match(docType, docTags, docExt) {
@@ -90,8 +91,8 @@ func (ss *ScenarioSet) SummaryFor(docType cbc.Key, docTags []cbc.Key, docExt []E
 			for k, v := range s.Codes {
 				summary.Codes[k] = v
 			}
-			for k, v := range s.Meta {
-				summary.Meta[k] = v
+			for k, v := range s.Ext {
+				summary.Ext[k] = v
 			}
 		}
 	}
@@ -99,9 +100,11 @@ func (ss *ScenarioSet) SummaryFor(docType cbc.Key, docTags []cbc.Key, docExt []E
 }
 
 func (ss *ScenarioSummary) addNote(note *cbc.Note) {
-	for _, n := range ss.Notes {
-		if n.Equals(note) {
-			return // already added
+	for i, n := range ss.Notes {
+		if n.SameAs(note) {
+			// replace
+			ss.Notes[i] = note
+			return
 		}
 	}
 	ss.Notes = append(ss.Notes, note)
@@ -173,7 +176,7 @@ func (s *Scenario) ValidateWithContext(ctx context.Context) error {
 		validation.Field(&s.Name),
 		validation.Field(&s.Note),
 		validation.Field(&s.Codes),
-		validation.Field(&s.Meta),
+		validation.Field(&s.Ext),
 	)
 	return err
 }