+++ tax/set_test.go
@@ -152,7 +152,7 @@ func TestSetValidation(t *testing.T) {
 					},
 				},
 			},
-			err: "0: ext: (foo: invalid.)",
+			err: "0: (ext: (foo: invalid.).)",
 		},
 		{
 			desc: "category extension",