+++ CHANGELOG.md
@@ -4,6 +4,40 @@ All notable changes to GOBL will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/). See also the [GOBL versions](https://docs.gobl.org/overview/versions) documentation site for more details.
 
+## [Unreleased]
+
+Significant set of changes around Scenario handling. Scenarios defined by tax regimes can now set tax extensions at the document level automatically. The objective here is to move away from external projects using scenario summaries directly, and instead use the absolute values set in the document.
+
+For example, the document format and type in Italy are now set inside the extensions and can be overriden if needed manually. This will be especially important when receiving and converting invoices into GOBL from external formats; its much easier to set specific values than trying to determine the appropriate tags.
+
+### Changed
+
+- `pt`: VAT rate key is now optional if `pt-saft-tax-rate` is provided.
+- `gr`: simplified validation to use tax categories.
+- `it`: always add `it-sdi-fiscal-regime` to Invoice suppliers.
+- `it`: renamed extension `it-sdi-retained-tax` to `it-sdi-retained`, now with validation on retained taxes.
+- `it`: renamed extension `it-sdi-natura` to `it-sdi-exempt`.
+- `bill.Invoice`: deprecated the `ScenarioSummary` method, as tax regimes themselves should be using extensions to apply all the correct data to a document up front.
+- `mx`: scenarios will now copy the document and relation types to the tax extensions.
+
+### Added
+
+- `tax.Category`: added `Validation` method support for custom validation of a tax combo for a specific tax category.
+- `tax.Scenario`: added "extensions" to be able to automatically update document level extensions based on the scenario detected.
+- `it`: added `ExtKeySDIDocumentType` as an extension that will be automatically included according to the scenario.
+- `it`: now adding `ExtKeySDIFormat` value to document instead of just referencing from scenarios.
+- `cbc.Note`: now provides `SameAs` method that will compare key attributes, but not the text payload. This is now used in Schema Summaries.
+- `bill.Line`: added `RequireLineTaxCategory` validation helper method.
+
+### Removed
+
+- `tax.Category`: removed `RateRequired` flag, regimes should instead should help users determine valid extensions (eg. PT and GR).
+
+### Fixed
+
+- `tax.Scenario`: potential issue around matching notes.
+- `tax.Set`: improved validation embedded error handling.
+
 ## [v0.114.0]
 
 ### Changed